backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
pdfminer.six==20221105
python-docx==1.1.0
supabase==2.4.3
openai==1.23.3
google-generativeai==0.5.4
httpx==0.27.0
cryptography==42.0.5
